Adjusts Button Defaults
Changes the default for the "Github" button, the "Forum" button, and the
"Rules" button. This doesn't modify the config files at all and so long
as those are done properly the config file will be used, however if they
break for some reason this will be used as the default instead.